Fix after review - rename, merge files.
