--- ipc/logger/interfaces.go
@@ -1,10 +1,0 @@
-package logger
-
-// Logger is the logger interface
-type Logger interface {
-	Trace(message string, args ...interface{})
-	Debug(message string, args ...interface{})
-	Info(message string, args ...interface{})
-	Warn(message string, args ...interface{})
-	Error(message string, args ...interface{})
-}
